microservices/ms1_flask/requirements.txt
flask
flasgger
flask-cors
mysql-connector-python
